Lazily init the loader in DSL
Lazily initialize the loader in the DSL so you can always use it regardless
of when the DSL is loaded.  This is required because to get it loading by
default in Titan you have to load it before the graph is available.

This change will slap the loader into the graph when it is first used
and reuse it from then on.